[MIG] l10n_ch_payment_slip: Migration to 11.0
Update headers
Port code to python3
Update README
Fix tests with correct report render calls
Adapt code to report module split
