+++ l10n_ch_payment_slip/__manifest__.py
@@ -1,17 +1,14 @@
-# -*- coding: utf-8 -*-
-# © 2012-2016 Camptocamp SA
+# Copyright 2012-2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 {'name': 'Switzerland - Payment Slip (BVR/ESR)',
  'summary': 'Print ESR/BVR payment slip with your invoices',
- 'version': '10.0.1.1.0',
+ 'version': '11.0.1.0.0',
  'author': "Camptocamp,Odoo Community Association (OCA)",
  'category': 'Localization',
  'website': 'http://www.camptocamp.com',
  'license': 'AGPL-3',
  'depends': [
-     'base',
-     'account',
-     'report',
+     'account_invoicing',
      'l10n_ch_base_bank',
      'base_transaction_id',  # OCA/bank-statement-reconcile
  ],
@@ -27,6 +24,6 @@
  'demo': [],
  'test': [],
  'auto_install': False,
- 'installable': False,
+ 'installable': True,
  'images': []
  }
