+++ l10n_ch_payment_slip/models/company.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# © 2012-2015 Camptocamp SA
+# Copyright 2012-2015 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from odoo import models, fields
