+++ l10n_ch_payment_slip/tests/test_v11_import.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# © 2014-2016 Camptocamp SA
+# Copyright 2014-2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 import base64
 from odoo.modules import get_module_resource
