+++ l10n_ch_payment_slip/wizard/bvr_batch_print.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
@@ -30,7 +29,7 @@ class BvrBatchPrintWizard(models.TransientModel):
     def check_generatable(self, invoices):
         try:
             invoices._check_bvr_generatable()
-        except UserError, e:
+        except UserError as e:
             return e.name
 
     @api.multi
