+++ l10n_ch_payment_slip/wizard/bvr_import.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# © 2014-2016 Camptocamp SA
+# Copyright 2014-2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 import base64
 import time
